tutorial notebook and some minor improvements
