+++ qubovert/utils/__init__.py
@@ -25,6 +25,7 @@ from ._approximate_extrema import *
 from ._hash import *
 from ._subgraph import *
 from ._normalize import *
+from ._values import *
 from ._solve_bruteforce import *
 from ._dict_arithmetic import *
 from ._pubomatrix import *
@@ -40,6 +41,7 @@ from ._approximate_extrema import __all__ as __all_ae__
 from ._hash import __all__ as __all_hash__
 from ._subgraph import __all__ as __all_subgraph__
 from ._normalize import __all__ as __all_normalize__
+from ._values import __all__ as __all_values__
 from ._solve_bruteforce import __all__ as __all_solve_bruteforce__
 from ._dict_arithmetic import __all__ as __all_dict_arithmetic__
 from ._pubomatrix import __all__ as __all_pubomatrix__
@@ -57,6 +59,7 @@ __all__ = (
     __all_hash__ +
     __all_subgraph__ +
     __all_normalize__ +
+    __all_values__ +
     __all_solve_bruteforce__ +
     __all_dict_arithmetic__ +
     __all_pubomatrix__ +
@@ -68,8 +71,8 @@ __all__ = (
 )
 
 del __all_warn__, __all_bh__, __all_ae__, __all_hash__, __all_subgraph__
-del __all_normalize__, __all_solve_bruteforce__, __all_dict_arithmetic__
-del __all_pubomatrix__, __all_pusomatrix__
+del __all_normalize__, __all_values__, __all_solve_bruteforce__
+del __all_dict_arithmetic__, __all_pubomatrix__, __all_pusomatrix__
 del __all_qubomatrix__, __all_qusomatrix__
 del __all_conversions__, __all_bo__
