+++ qubovert/utils/_solve_bruteforce.py
@@ -14,327 +14,17 @@
 
 """_solve_bruteforce.py.
 
-This file contains bruteforce solvers for QUBO/PUBO and QUSO/PUSO, as well
-as QUBO/PUBO and PUSO objective function evaluators.
+This file contains bruteforce solvers for QUBO/PUBO and QUSO/PUSO.
 
 """
 
+import itertools
+from . import pubo_value, puso_value
+
 __all__ = (
-    'boolean_to_spin', 'spin_to_boolean',
-    'decimal_to_spin', 'spin_to_decimal',
-    'decimal_to_boolean', 'boolean_to_decimal',
-    'pubo_value', 'qubo_value', 'puso_value', 'quso_value',
     'solve_pubo_bruteforce', 'solve_qubo_bruteforce',
     'solve_puso_bruteforce', 'solve_quso_bruteforce'
 )
-
-
-def boolean_to_spin(x):
-    """boolean_to_spin.
-
-    Convert a boolean number in {0, 1} to a spin in {1, -1}, in that order.
-
-    Parameters
-    ----------
-    x : int, iterable of ints, or dict mapping labels to ints.
-        Each integer is either 0 or 1.
-
-    Returns
-    -------
-    z : int, iterable of ints, or dict mapping labels to ints.
-        Each integer is either 1 or -1.
-
-    Example
-    -------
-    >>> boolean_to_spin(0)  # will print 1
-    >>> boolean_to_spin(1)  # will print -1
-    >>> boolean_to_spin([0, 1, 1])  # will print [1, -1, -1]
-    >>> boolean_to_spin({"a": 0, "b": 1})  # will print {"a": 1, "b": -1}
-
-    """
-    convert = {0: 1, 1: -1}
-    if isinstance(x, (int, float)) and x in convert:
-        return convert[x]
-    elif isinstance(x, dict):
-        return {k: convert[v] for k, v in x.items()}
-    return type(x)(convert[i] for i in x)
-
-
-def spin_to_boolean(z):
-    """spin_to_boolean.
-
-    Convert a spin in {1, -1} to a boolean variable in {0, 1}, in that order.
-
-    Parameters
-    ----------
-    z : int, iterable of ints, or dict mapping labels to ints.
-        Each integer is either 1 or -1.
-
-    Returns
-    -------
-    x : int, iterable of ints, or dict mapping labels to ints.
-        Each integer is either 0 or 1.
-
-    Example
-    -------
-    >>> spin_to_boolean(-1)  # will print 1
-    >>> spin_to_boolean(1)  # will print 0
-    >>> spin_to_boolean([-1, 1, 1])  # will print [1, 0, 0]
-    >>> spin_to_boolean({"a": -1, "b": 1})  # will print {"a": 1, "b": 0}
-
-    """
-    convert = {-1: 1, 1: 0}
-    if isinstance(z, (int, float)) and z in convert:
-        return convert[z]
-    elif isinstance(z, dict):
-        return {k: convert[v] for k, v in z.items()}
-    return type(z)(convert[i] for i in z)
-
-
-def decimal_to_boolean(d, num_bits=None):
-    """decimal_to_boolean.
-
-    Convert the integer ``d`` to its boolean representation.
-
-    Parameters
-    ----------
-    d : int >= 0.
-        Number to convert to binary.
-    num_bits : int >= 0 (optional, defaults to None).
-        Number of bits in the representation. If ``num_bits is None``, then
-        the minimum number of bits required will be used.
-
-    Returns
-    -------
-    b : tuple of length ``num_bits``.
-        Each element of ``b`` is a 0 or 1.
-
-    Example
-    -------
-    >>> decimal_to_boolean(10, 7)
-    (0, 0, 0, 1, 0, 1, 0)
-
-    >>> decimal_to_boolean(10)
-    (1, 0, 1, 0)
-
-    """
-    if int(d) != d or d < 0:
-        raise ValueError("``d`` must be an integer >- 0.")
-    b = bin(d)[2:]
-    lb = len(b)
-    if num_bits is None:
-        num_bits = lb
-    elif num_bits < lb:
-        raise ValueError("Not enough bits to represent the number.")
-    return (0,) * (num_bits - lb) + tuple(int(x) for x in b)
-
-
-def boolean_to_decimal(b):
-    """boolean_to_decimal.
-
-    Convert a bit string to its decimal form.
-
-    Parameters
-    ----------
-    b : tuple or list of 0s and 1s.
-        The binary bit string.
-
-    Returns
-    -------
-    d : int.
-
-    Examples
-    --------
-    >>> boolean_to_decimal((1, 1, 0))
-    6
-
-    """
-    return int("".join(str(x) for x in b), base=2) if b else 0
-
-
-def decimal_to_spin(d, num_spins=None):
-    """decimal_to_spin.
-
-    Convert the integer ``d`` to its spin representation (ie its binary
-    representation, but with 1 and -1 instead of 0 and 1).
-
-    Parameters
-    ----------
-    d : int >= 0.
-        Number to convert to binary.
-    num_spins : int >= 0 (optional, defaults to None).
-        Number of bits in the representation. If ``num_spins is None``, then
-        the minimum number of bits required will be used.
-
-    Returns
-    -------
-    b : tuple of length ``num_spins``.
-        Each element of ``b`` is a 0 or 1.
-
-    Example
-    -------
-    >>> decimal_to_spin(10, 7)
-    (1, 1, 1, -1, 1, -1, 1)
-
-    >>> decimal_to_spin(10)
-    (-1, 1, -1, 1)
-
-    """
-    return boolean_to_spin(decimal_to_boolean(d, num_spins))
-
-
-def spin_to_decimal(b):
-    """spin_to_decimal.
-
-    Convert a spin string to its decimal form.
-
-    Parameters
-    ----------
-    b : tuple or list of 1s and -1s.
-        The spin bit string.
-
-    Returns
-    -------
-    d : int.
-
-    Examples
-    --------
-    >>> spin_to_decimal((-1, -1, 1))
-    6
-
-    """
-    return boolean_to_decimal(spin_to_boolean(b))
-
-
-def pubo_value(x, P):
-    r"""pubo_value.
-
-    Find the value of
-    :math:`\sum_{i,...,j} P_{i...j} x_{i} ... x_{j}`
-
-    Parameters
-    ----------
-    x : dict or iterable.
-        Maps boolean variable indices to their boolean values, 0 or 1. Ie
-        ``x[i]`` must be the boolean value of variable i.
-    P : dict, qubovert.utils.PUBOMatrix, or qubovert.PUBO object.
-        Maps tuples of boolean variables indices to the P value.
-
-    Returns
-    -------
-    value : float.
-        The value of the PUBO with the given assignment `x`. Ie
-
-    Example
-    -------
-    >>> P = {(0, 0): 1, (0, 1): -1}
-    >>> x = {0: 1, 1: 0}
-    >>> pubo_value(x, P)
-    1
-
-    """
-    return sum(v for k, v in P.items() if all(x[i] for i in k))
-
-
-def qubo_value(x, Q):
-    r"""qubo_value.
-
-    Find the value of
-    :math:`\sum_{i,j} Q_{ij} x_{i} x_{j}`
-
-    Parameters
-    ----------
-    x : dict or iterable.
-        Maps boolean variable indices to their boolean values, 0 or 1. Ie
-        ``x[i]`` must be the boolean value of variable i.
-    Q : dict or qubovert.utils.QUBOMatrix object.
-        Maps tuples of boolean variables indices to the Q value.
-
-    Returns
-    -------
-    value : float.
-        The value of the QUBO with the given assignment `x`. Ie
-
-        >>> sum(
-                Q[(i, j)] * x[i] * x[j]
-                for i in range(n) for j in range(n)
-            )
-
-    Example
-    -------
-    >>> Q = {(0, 0): 1, (0, 1): -1}
-    >>> x = {0: 1, 1: 0}
-    >>> qubo_value(x, Q)
-    1
-
-    """
-    return pubo_value(x, Q)
-
-
-def puso_value(z, H):
-    r"""puso_value.
-
-    Find the value of
-        :math:`\sum_{i,...,j} H_{i...j} z_{i} ... z_{j}`.
-
-    Parameters
-    ----------
-    z: dict or iterable.
-        Maps variable labels to their values, 1 or -1. Ie z[i] must be the
-        value of variable i.
-    H : dict, qubovert.utils.PUSOMatrix, or qubovert.PUSO object.
-        Maps spin labels to values.
-
-    Returns
-    -------
-    value : float.
-        The value of the PUSO with the given assignment `z`.
-
-    Example
-    -------
-    >>> H = {(0, 1): -1, (0,): 1}
-    >>> z = {0: -1, 1: 1}
-    >>> puso_value(z, H)
-    0
-
-    """
-    return sum(
-        v * pow(-1, [z[i] for i in k].count(-1) % 2)
-        for k, v in H.items()
-    )
-
-
-def quso_value(z, L):
-    r"""quso_value.
-
-    Find the value of
-        :math:`\sum_{i,j} J_{ij} z_{i} z_{j} + \sum_{i} h_{i} z_{i}`.
-
-    The J's are encoded by keys with pairs of labels in L, and the h's are
-    encoded by keys with a single label in L.
-
-    Parameters
-    ----------
-    z: dict or iterable.
-        Maps variable labels to their values, 1 or -1. Ie z[i] must be the
-        value of variable i.
-    L : dict, qubovert.utils.QUSOMatrix, or qubovert.QUSO object.
-        Maps spin labels to values.
-
-    Returns
-    -------
-    value : float.
-        The value of the QUSO with the given assignment `z`.
-
-    Example
-    -------
-    >>> L = {(0, 1): -1, (0,): 1}
-    >>> z = {0: -1, 1: 1}
-    >>> quso_value(z, L)
-    0
-
-    """
-    return puso_value(z, L)
 
 
 def _solve_bruteforce(D, all_solutions, valid, spin):
@@ -407,10 +97,8 @@ def _solve_bruteforce(D, all_solutions, valid, spin):
 
     best = None, {}
     all_sols = {None: [{}]}
-    convert = decimal_to_spin if spin else decimal_to_boolean
 
-    for n in range(1 << N):
-        test_sol = convert(n, N)
+    for test_sol in itertools.product((1, -1) if spin else (0, 1), repeat=N):
         x = {mapping[i]: v for i, v in enumerate(test_sol)}
         if not valid(x):
             continue
