+++ tests/test_qubo.py
@@ -31,14 +31,13 @@ def test_pretty_str():
 
     def equal(expression, string):
         assert expression.pretty_str() == string
-        assert QUBO.pretty_str(dict(expression)) == string
 
     x = [QUBO() + {(i,): 1} for i in range(3)]
     a, b = Symbol('a'), Symbol('b')
 
     equal(x[0], "x(0)")
     equal(-x[0], "-x(0)")
-    equal(x[0] * 0, "")
+    equal(x[0] * 0, "0")
     equal(2*x[0]*x[1] - 3*x[2], "2 x(0) x(1) - 3 x(2)")
     equal(0*x[0] + 1, "1")
     equal(0*x[0] - 1, "-1")
@@ -48,6 +47,19 @@ def test_pretty_str():
     equal(2*x[0]*x[1] - x[2], "2 x(0) x(1) - x(2)")
     equal(-x[2] + x[0]*x[1], "-x(2) + x(0) x(1)")
     equal(-2*x[2] + 2*x[0]*x[1], "-2 x(2) + 2 x(0) x(1)")
+
+
+def test_create_var():
+
+    d = QUBO.create_var(0)
+    assert d == {(0,): 1}
+    assert d.name == 0
+    assert type(d) == QUBO
+
+    d = QUBO.create_var('x')
+    assert d == {('x',): 1}
+    assert d.name == 'x'
+    assert type(d) == QUBO
 
 
 problem = QUBO({('a',): -1, ('b',): 2, ('a', 'b'): -3, ('b', 'c'): -4, (): -2})
