+++ tests/utils/test_pubomatrix.py
@@ -26,14 +26,13 @@ def test_pretty_str():
 
     def equal(expression, string):
         assert expression.pretty_str() == string
-        assert PUBOMatrix.pretty_str(dict(expression)) == string
 
     x = [PUBOMatrix() + {(i,): 1} for i in range(3)]
     a, b = Symbol('a'), Symbol('b')
 
+    equal(x[0] * 0, "0")
     equal(x[0], "x(0)")
     equal(-x[0], "-x(0)")
-    equal(x[0] * 0, "")
     equal(2*x[0]*x[1] - 3*x[2], "2 x(0) x(1) - 3 x(2)")
     equal(0*x[0] + 1, "1")
     equal(0*x[0] - 1, "-1")
@@ -44,8 +43,18 @@ def test_pretty_str():
     equal(-x[2] + x[0]*x[1], "-x(2) + x(0) x(1)")
     equal(-2*x[2] + 2*x[0]*x[1], "-2 x(2) + 2 x(0) x(1)")
 
-    # test when there is a zero element
-    assert PUBOMatrix.pretty_str({(0,): 1, (1, 2): 0}, 'y') == "y(0)"
+
+def test_create_var():
+
+    d = PUBOMatrix.create_var(0)
+    assert d == {(0,): 1}
+    assert d.name == 0
+    assert type(d) == PUBOMatrix
+
+    d = PUBOMatrix.create_var(1)
+    assert d == {(1,): 1}
+    assert d.name == 1
+    assert type(d) == PUBOMatrix
 
 
 def test_pubo_checkkey():
